Split git.py into multiple files
Prepares for reworking of the handlers and also makesit easier to get an
overview of the file
