+++ backend/infrahub/git/__init__.py
@@ -1,0 +1,29 @@
+from infrahub.git.handlers import (
+    handle_git_check_message,
+    handle_git_rpc_message,
+    handle_git_transform_message,
+)
+from infrahub.git.repository import (
+    BRANCHES_DIRECTORY_NAME,
+    COMMITS_DIRECTORY_NAME,
+    TEMPORARY_DIRECTORY_NAME,
+    InfrahubRepository,
+    RepoFileInformation,
+    Worktree,
+    extract_repo_file_information,
+    initialize_repositories_directory,
+)
+
+__all__ = [
+    "BRANCHES_DIRECTORY_NAME",
+    "COMMITS_DIRECTORY_NAME",
+    "TEMPORARY_DIRECTORY_NAME",
+    "InfrahubRepository",
+    "RepoFileInformation",
+    "Worktree",
+    "extract_repo_file_information",
+    "initialize_repositories_directory",
+    "handle_git_check_message",
+    "handle_git_rpc_message",
+    "handle_git_transform_message",
+]
